HS-31-03 (1/2): extract PluginArtifact/Project/Activity repos
Move the remaining three domains out of MeetingDatabase into the package:
- plugins.py: PluginArtifactRepository (intent windows, plugin runs/jobs, artifacts)
- projects.py: ProjectRepository (projects, associations, detection log)
- activity.py: ActivityRepository (local activity-intelligence ledger)

core.py is now a thin container (1224 lines, from 5481): __init__, _connection,
_ensure_schema, _apply_schema, schema SQL, and the get_database singleton — no
domain methods remain.

AST-based extraction (helpers were scattered) with a coverage assertion. The one
inter-repo call (activity rule -> get_project) uses the self._db back-reference.
~469 call sites moved to db.{plugins,projects,activity}.*; fake doubles got the
matching repo self-properties. Suite green at 2062 / 14 skipped; ruff-clean.

The MeetingDatabase name is retired to `Database` in the follow-up commit.
