+++ tests/integration/test_mermaid_architecture_pipeline.py
@@ -105,7 +105,7 @@ def test_mermaid_architecture_plugin_lands_diagram_artifact(temp_db) -> None:
     assert executed.output["confidence_hint"] == 1.0
 
     # Step 3 — persist the run, then synthesize + persist the artifact.
-    temp_db.record_plugin_run(
+    temp_db.plugins.record_plugin_run(
         meeting_id="m-mermaid-it",
         window_id="w-1",
         plugin_id="mermaid_architecture",
@@ -130,6 +130,6 @@ def test_mermaid_architecture_plugin_lands_diagram_artifact(temp_db) -> None:
     assert lineages[0].artifact_id == diagram.artifact_id
 
     # Persisted in the DB.
-    persisted = temp_db.list_artifacts("m-mermaid-it")
+    persisted = temp_db.plugins.list_artifacts("m-mermaid-it")
     assert {a.id for a in persisted} == {diagram.artifact_id}
     assert {a.artifact_type for a in persisted} == {"diagram"}
