+++ tests/unit/test_activity_candidates.py
@@ -10,21 +10,21 @@ from holdspeak.db import MeetingDatabase
 
 def test_preview_calendar_meeting_candidates_uses_local_activity_records(tmp_path):
     db = MeetingDatabase(tmp_path / "holdspeak.db")
-    outlook = db.upsert_activity_record(
+    outlook = db.activity.upsert_activity_record(
         source_browser="safari",
         url="https://outlook.office.com/calendar/item/123",
         title="Customer planning meeting",
         domain="outlook.office.com",
         last_seen_at=datetime(2026, 4, 27, 9, 0, 0),
     )
-    db.upsert_activity_record(
+    db.activity.upsert_activity_record(
         source_browser="safari",
         url="https://example.com/not-calendar",
         title="Regular page",
         domain="example.com",
     )
 
-    previews = preview_calendar_meeting_candidates(db.list_activity_records(limit=10))
+    previews = preview_calendar_meeting_candidates(db.activity.list_activity_records(limit=10))
 
     assert len(previews) == 1
     assert previews[0].title == "Customer planning meeting"
@@ -36,7 +36,7 @@ def test_preview_calendar_meeting_candidates_uses_local_activity_records(tmp_pat
 
 def test_preview_calendar_meeting_candidates_falls_back_to_domain_titles(tmp_path):
     db = MeetingDatabase(tmp_path / "holdspeak.db")
-    record = db.upsert_activity_record(
+    record = db.activity.upsert_activity_record(
         source_browser="firefox",
         url="https://meet.google.com/abc-defg-hij",
         title=None,
@@ -51,7 +51,7 @@ def test_preview_calendar_meeting_candidates_falls_back_to_domain_titles(tmp_pat
 
 def test_preview_calendar_meeting_candidates_extracts_visible_time_hints(tmp_path):
     db = MeetingDatabase(tmp_path / "holdspeak.db")
-    record = db.upsert_activity_record(
+    record = db.activity.upsert_activity_record(
         source_browser="safari",
         url="https://outlook.office.com/calendar/item/123",
         title="Customer sync 2026-04-27 15:00-15:30",
