+++ tests/unit/test_activity_context.py
@@ -23,7 +23,7 @@ class _CapturePlugin:
 
 def test_build_activity_context_serializes_recent_records(tmp_path):
     db = MeetingDatabase(tmp_path / "holdspeak.db")
-    db.upsert_activity_record(
+    db.activity.upsert_activity_record(
         source_browser="safari",
         source_profile="default",
         url="https://example.atlassian.net/browse/HS-805",
@@ -50,7 +50,7 @@ def test_activity_context_provider_can_refresh_once(tmp_path):
 
     def importer(**kwargs):
         calls.append(kwargs)
-        kwargs["db"].upsert_activity_record(
+        kwargs["db"].activity.upsert_activity_record(
             source_browser="firefox",
             url="https://miro.com/app/board/uXjVContext/",
             title="Context board",
@@ -72,7 +72,7 @@ def test_activity_context_provider_can_refresh_once(tmp_path):
 
 def test_plugin_host_injects_activity_context_for_any_plugin(tmp_path):
     db = MeetingDatabase(tmp_path / "holdspeak.db")
-    db.upsert_activity_record(
+    db.activity.upsert_activity_record(
         source_browser="safari",
         url="https://github.com/openai/codex/pull/12",
         title="PR 12",
