+++ tests/unit/test_db.py
@@ -678,7 +678,7 @@ class TestActivityLedgerPersistence:
     def test_upsert_activity_record_persists_locally(self, db):
         seen = datetime(2026, 4, 26, 9, 30, 0)
 
-        record = db.upsert_activity_record(
+        record = db.activity.upsert_activity_record(
             source_browser="Safari",
             source_profile="default",
             source_path_hash="hash-safari",
@@ -700,7 +700,7 @@ class TestActivityLedgerPersistence:
         assert record.last_visit_raw == "799320600.0"
         assert record.last_seen_at == seen
 
-        listed = db.list_activity_records(source_browser="safari")
+        listed = db.activity.list_activity_records(source_browser="safari")
         assert len(listed) == 1
         assert listed[0].id == record.id
 
@@ -708,7 +708,7 @@ class TestActivityLedgerPersistence:
         first = datetime(2026, 4, 26, 9, 0, 0)
         last = datetime(2026, 4, 26, 11, 0, 0)
 
-        created = db.upsert_activity_record(
+        created = db.activity.upsert_activity_record(
             source_browser="firefox",
             source_profile="work",
             url="https://miro.com/app/board/uXjVMiro123/?utm_source=email",
@@ -718,7 +718,7 @@ class TestActivityLedgerPersistence:
             last_seen_at=last,
             last_visit_raw="1745665200000000",
         )
-        merged = db.upsert_activity_record(
+        merged = db.activity.upsert_activity_record(
             source_browser="firefox",
             source_profile="work",
             url="https://miro.com/app/board/uXjVMiro123?utm_source=email",
@@ -734,13 +734,13 @@ class TestActivityLedgerPersistence:
         assert merged.visit_count == 5
         assert merged.first_seen_at == first
         assert merged.last_seen_at == last
-        assert db.list_activity_records(source_browser="firefox", source_profile="work") == [merged]
+        assert db.activity.list_activity_records(source_browser="firefox", source_profile="work") == [merged]
 
     def test_duplicate_activity_records_merge_by_entity(self, db):
         first = datetime(2026, 4, 26, 9, 0, 0)
         second = datetime(2026, 4, 26, 10, 0, 0)
 
-        created = db.upsert_activity_record(
+        created = db.activity.upsert_activity_record(
             source_browser="safari",
             url="https://github.com/acme/app/pull/42",
             title="PR 42",
@@ -749,7 +749,7 @@ class TestActivityLedgerPersistence:
             entity_type="github_pull_request",
             entity_id="acme/app#42",
         )
-        merged = db.upsert_activity_record(
+        merged = db.activity.upsert_activity_record(
             source_browser="safari",
             url="https://github.com/acme/app/pull/42/files",
             title="PR 42 files",
@@ -763,12 +763,12 @@ class TestActivityLedgerPersistence:
         assert merged.normalized_url == "https://github.com/acme/app/pull/42/files"
         assert merged.first_seen_at == first
         assert merged.last_seen_at == second
-        assert len(db.list_activity_records(entity_type="github_pull_request")) == 1
+        assert len(db.activity.list_activity_records(entity_type="github_pull_request")) == 1
 
     def test_activity_import_checkpoint_round_trips_per_source_profile(self, db):
         imported_at = datetime(2026, 4, 26, 12, 0, 0)
 
-        checkpoint = db.set_activity_import_checkpoint(
+        checkpoint = db.activity.set_activity_import_checkpoint(
             source_browser="Safari",
             source_profile="default",
             source_path_hash="path-hash",
@@ -785,7 +785,7 @@ class TestActivityLedgerPersistence:
         assert checkpoint.last_imported_at == imported_at
         assert checkpoint.enabled is True
 
-        db.set_activity_import_checkpoint(
+        db.activity.set_activity_import_checkpoint(
             source_browser="safari",
             source_profile="default",
             source_path_hash="path-hash",
@@ -794,7 +794,7 @@ class TestActivityLedgerPersistence:
             last_error="temporary lock",
             enabled=False,
         )
-        updated = db.get_activity_import_checkpoint(
+        updated = db.activity.get_activity_import_checkpoint(
             source_browser="safari",
             source_profile="default",
             source_path_hash="path-hash",
@@ -807,33 +807,33 @@ class TestActivityLedgerPersistence:
     def test_delete_activity_records_supports_retention_filters(self, db):
         old = datetime(2026, 4, 1, 9, 0, 0)
         recent = datetime(2026, 4, 26, 9, 0, 0)
-        db.upsert_activity_record(
+        db.activity.upsert_activity_record(
             source_browser="safari",
             url="https://old.example.com/ticket",
             domain="old.example.com",
             last_seen_at=old,
         )
-        db.upsert_activity_record(
+        db.activity.upsert_activity_record(
             source_browser="safari",
             url="https://recent.example.com/ticket",
             domain="recent.example.com",
             last_seen_at=recent,
         )
 
-        deleted = db.delete_activity_records(older_than=datetime(2026, 4, 10))
+        deleted = db.activity.delete_activity_records(older_than=datetime(2026, 4, 10))
 
         assert deleted == 1
-        remaining = db.list_activity_records()
+        remaining = db.activity.list_activity_records()
         assert len(remaining) == 1
         assert remaining[0].domain == "recent.example.com"
 
     def test_activity_privacy_settings_default_enabled_and_updateable(self, db):
-        settings = db.get_activity_privacy_settings()
+        settings = db.activity.get_activity_privacy_settings()
         assert settings["enabled"] is True
         assert settings["paused"] is False
         assert settings["retention_days"] == 30
 
-        updated = db.update_activity_privacy_settings(
+        updated = db.activity.update_activity_privacy_settings(
             enabled=False,
             retention_days=14,
         )
@@ -843,21 +843,21 @@ class TestActivityLedgerPersistence:
         assert updated["retention_days"] == 14
 
     def test_activity_domain_rules_match_subdomains(self, db):
-        rule = db.upsert_activity_domain_rule(domain="Example.COM", action="exclude")
+        rule = db.activity.upsert_activity_domain_rule(domain="Example.COM", action="exclude")
 
         assert rule["domain"] == "example.com"
-        assert db.is_activity_domain_excluded("example.com") is True
-        assert db.is_activity_domain_excluded("docs.example.com") is True
-        assert db.is_activity_domain_excluded("other.com") is False
+        assert db.activity.is_activity_domain_excluded("example.com") is True
+        assert db.activity.is_activity_domain_excluded("docs.example.com") is True
+        assert db.activity.is_activity_domain_excluded("other.com") is False
 
-        assert db.delete_activity_domain_rule("example.com") is True
-        assert db.is_activity_domain_excluded("example.com") is False
+        assert db.activity.delete_activity_domain_rule("example.com") is True
+        assert db.activity.is_activity_domain_excluded("example.com") is False
 
     def test_activity_project_rules_preview_and_apply_records(self, db):
-        db.create_project(project_id="holdspeak", name="HoldSpeak")
-        db.create_project(project_id="other", name="Other")
+        db.projects.create_project(project_id="holdspeak", name="HoldSpeak")
+        db.projects.create_project(project_id="other", name="Other")
         first_seen = datetime(2026, 4, 26, 9, 0, 0)
-        db.upsert_activity_record(
+        db.activity.upsert_activity_record(
             source_browser="safari",
             url="https://example.atlassian.net/browse/HS-123",
             title="HS-123 activity mapping",
@@ -866,7 +866,7 @@ class TestActivityLedgerPersistence:
             entity_type="jira_ticket",
             entity_id="HS-123",
         )
-        db.upsert_activity_record(
+        db.activity.upsert_activity_record(
             source_browser="safari",
             url="https://example.atlassian.net/browse/OTHER-1",
             title="OTHER-1 activity mapping",
@@ -876,14 +876,14 @@ class TestActivityLedgerPersistence:
             entity_id="OTHER-1",
         )
 
-        low_priority = db.create_activity_project_rule(
+        low_priority = db.activity.create_activity_project_rule(
             project_id="other",
             name="All Jira",
             match_type="entity_type",
             pattern="jira_ticket",
             priority=100,
         )
-        high_priority = db.create_activity_project_rule(
+        high_priority = db.activity.create_activity_project_rule(
             project_id="holdspeak",
             name="HoldSpeak tickets",
             match_type="entity_id_prefix",
@@ -894,8 +894,8 @@ class TestActivityLedgerPersistence:
 
         assert isinstance(high_priority, ActivityProjectRule)
         assert high_priority.project_name == "HoldSpeak"
-        assert db.list_activity_project_rules() == [high_priority, low_priority]
-        preview = db.preview_activity_project_rule(
+        assert db.activity.list_activity_project_rules() == [high_priority, low_priority]
+        preview = db.activity.preview_activity_project_rule(
             project_id="holdspeak",
             match_type="entity_id_prefix",
             pattern="HS-",
@@ -903,22 +903,22 @@ class TestActivityLedgerPersistence:
         )
         assert [record.entity_id for record in preview] == ["HS-123"]
 
-        assert db.apply_activity_project_rules() == 2
-        records = db.list_activity_records(limit=10)
+        assert db.activity.apply_activity_project_rules() == 2
+        records = db.activity.list_activity_records(limit=10)
         assert {record.entity_id: record.project_id for record in records} == {
             "HS-123": "holdspeak",
             "OTHER-1": "other",
         }
 
     def test_activity_project_rules_update_disable_and_delete(self, db):
-        db.create_project(project_id="holdspeak", name="HoldSpeak")
-        rule = db.create_activity_project_rule(
+        db.projects.create_project(project_id="holdspeak", name="HoldSpeak")
+        rule = db.activity.create_activity_project_rule(
             project_id="holdspeak",
             match_type="domain",
             pattern="Example.COM",
         )
 
-        updated = db.update_activity_project_rule(
+        updated = db.activity.update_activity_project_rule(
             rule.id,
             name="Example",
             enabled=False,
@@ -929,15 +929,15 @@ class TestActivityLedgerPersistence:
         assert updated.name == "Example"
         assert updated.enabled is False
         assert updated.priority == 300
-        assert db.list_activity_project_rules() == []
-        assert db.list_activity_project_rules(include_disabled=True) == [updated]
-        assert db.delete_activity_project_rule(rule.id) is True
-        assert db.list_activity_project_rules(include_disabled=True) == []
+        assert db.activity.list_activity_project_rules() == []
+        assert db.activity.list_activity_project_rules(include_disabled=True) == [updated]
+        assert db.activity.delete_activity_project_rule(rule.id) is True
+        assert db.activity.list_activity_project_rules(include_disabled=True) == []
 
     def test_activity_enrichment_connector_state_round_trips(self, db):
         run_at = datetime(2026, 4, 27, 10, 30, 0)
 
-        state = db.upsert_activity_enrichment_connector(
+        state = db.activity.upsert_activity_enrichment_connector(
             connector_id="gh",
             enabled=True,
             settings={"timeout_seconds": 4, "max_bytes": 2048},
@@ -950,7 +950,7 @@ class TestActivityLedgerPersistence:
         assert state.settings == {"timeout_seconds": 4, "max_bytes": 2048}
         assert state.last_error == "not run yet"
 
-        updated = db.record_activity_enrichment_run(
+        updated = db.activity.record_activity_enrichment_run(
             connector_id="gh",
             last_run_at=run_at,
         )
@@ -959,10 +959,10 @@ class TestActivityLedgerPersistence:
         assert updated.settings == {"timeout_seconds": 4, "max_bytes": 2048}
         assert updated.last_run_at == run_at
         assert updated.last_error is None
-        assert db.list_activity_enrichment_connectors() == [updated]
+        assert db.activity.list_activity_enrichment_connectors() == [updated]
 
     def test_activity_annotations_attach_to_records_and_delete_by_connector(self, db):
-        record = db.upsert_activity_record(
+        record = db.activity.upsert_activity_record(
             source_browser="safari",
             url="https://github.com/openai/codex/pull/42",
             title="PR 42",
@@ -971,7 +971,7 @@ class TestActivityLedgerPersistence:
             entity_id="openai/codex#42",
         )
 
-        annotation = db.create_activity_annotation(
+        annotation = db.activity.create_activity_annotation(
             activity_record_id=record.id,
             source_connector_id="gh",
             annotation_type="github_pr",
@@ -986,15 +986,15 @@ class TestActivityLedgerPersistence:
         assert annotation.annotation_type == "github_pr"
         assert annotation.value == {"state": "OPEN", "labels": ["activity"]}
         assert annotation.confidence == 1.0
-        assert db.list_activity_annotations(activity_record_id=record.id) == [annotation]
-        assert db.list_activity_annotations(source_connector_id="gh") == [annotation]
+        assert db.activity.list_activity_annotations(activity_record_id=record.id) == [annotation]
+        assert db.activity.list_activity_annotations(source_connector_id="gh") == [annotation]
 
-        assert db.delete_activity_annotations(source_connector_id="gh") == 1
-        assert db.list_activity_annotations(source_connector_id="gh") == []
+        assert db.activity.delete_activity_annotations(source_connector_id="gh") == 1
+        assert db.activity.list_activity_annotations(source_connector_id="gh") == []
 
     def test_activity_annotations_validate_record_reference(self, db):
         with pytest.raises(ValueError, match="activity record not found"):
-            db.create_activity_annotation(
+            db.activity.create_activity_annotation(
                 activity_record_id=999,
                 source_connector_id="gh",
                 annotation_type="github_pr",
@@ -1003,14 +1003,14 @@ class TestActivityLedgerPersistence:
     def test_activity_meeting_candidates_round_trip_and_status_update(self, db):
         starts_at = datetime(2026, 4, 27, 15, 0, 0)
         ends_at = datetime(2026, 4, 27, 15, 30, 0)
-        record = db.upsert_activity_record(
+        record = db.activity.upsert_activity_record(
             source_browser="safari",
             url="https://outlook.office.com/calendar/item/123",
             title="Customer sync",
             domain="outlook.office.com",
         )
 
-        candidate = db.create_activity_meeting_candidate(
+        candidate = db.activity.create_activity_meeting_candidate(
             source_connector_id="calendar_activity",
             source_activity_record_id=record.id,
             title="Customer sync",
@@ -1029,59 +1029,59 @@ class TestActivityLedgerPersistence:
         assert candidate.meeting_url == record.url
         assert candidate.confidence == 0.8
         assert candidate.status == "candidate"
-        assert db.list_activity_meeting_candidates(status="candidate") == [candidate]
+        assert db.activity.list_activity_meeting_candidates(status="candidate") == [candidate]
 
-        updated = db.update_activity_meeting_candidate_status(candidate.id, "armed")
+        updated = db.activity.update_activity_meeting_candidate_status(candidate.id, "armed")
         assert updated is not None
         assert updated.status == "armed"
-        assert db.list_activity_meeting_candidates(status="armed") == [updated]
-        assert db.get_activity_meeting_candidate(candidate.id) == updated
+        assert db.activity.list_activity_meeting_candidates(status="armed") == [updated]
+        assert db.activity.get_activity_meeting_candidate(candidate.id) == updated
 
-        started = db.mark_activity_meeting_candidate_started(
+        started = db.activity.mark_activity_meeting_candidate_started(
             candidate.id,
             meeting_id="meeting-123",
         )
         assert started is not None
         assert started.status == "started"
         assert started.started_meeting_id == "meeting-123"
-        assert db.list_activity_meeting_candidates(status="started") == [started]
+        assert db.activity.list_activity_meeting_candidates(status="started") == [started]
 
-        assert db.delete_activity_meeting_candidates(source_connector_id="calendar_activity") == 1
-        assert db.list_activity_meeting_candidates(source_connector_id="calendar_activity") == []
+        assert db.activity.delete_activity_meeting_candidates(source_connector_id="calendar_activity") == 1
+        assert db.activity.list_activity_meeting_candidates(source_connector_id="calendar_activity") == []
 
     def test_activity_meeting_candidates_validate_status_and_record_reference(self, db):
         with pytest.raises(ValueError, match="activity record not found"):
-            db.create_activity_meeting_candidate(
+            db.activity.create_activity_meeting_candidate(
                 source_connector_id="calendar_activity",
                 source_activity_record_id=999,
                 title="Missing source",
             )
 
         with pytest.raises(ValueError, match="candidate status"):
-            db.create_activity_meeting_candidate(
+            db.activity.create_activity_meeting_candidate(
                 source_connector_id="calendar_activity",
                 title="Bad status",
                 status="auto_record",
             )
 
     def test_activity_meeting_candidates_dedupe_repeated_saves(self, db):
-        record = db.upsert_activity_record(
+        record = db.activity.upsert_activity_record(
             source_browser="safari",
             url="https://outlook.office.com/calendar/item/123",
             title="Customer sync",
             domain="outlook.office.com",
         )
 
-        first = db.create_activity_meeting_candidate(
+        first = db.activity.create_activity_meeting_candidate(
             source_connector_id="calendar_activity",
             source_activity_record_id=record.id,
             title="Customer sync",
             meeting_url=record.url,
             confidence=0.55,
         )
-        armed = db.update_activity_meeting_candidate_status(first.id, "armed")
+        armed = db.activity.update_activity_meeting_candidate_status(first.id, "armed")
         assert armed is not None
-        second = db.create_activity_meeting_candidate(
+        second = db.activity.create_activity_meeting_candidate(
             source_connector_id="calendar_activity",
             source_activity_record_id=record.id,
             title="Customer sync updated",
@@ -1095,7 +1095,7 @@ class TestActivityLedgerPersistence:
         assert second.starts_at == datetime(2026, 4, 27, 15, 0, 0)
         assert second.confidence == 0.8
         assert second.status == "armed"
-        assert db.list_activity_meeting_candidates() == [second]
+        assert db.activity.list_activity_meeting_candidates() == [second]
 
 
 class TestDeferredIntelQueue:
@@ -1277,7 +1277,7 @@ class TestMirPersistence:
     def test_record_and_list_intent_windows(self, db, sample_meeting):
         db.meetings.save_meeting(sample_meeting)
 
-        db.record_intent_window(
+        db.plugins.record_intent_window(
             meeting_id=sample_meeting.id,
             window_id=f"{sample_meeting.id}:w0001",
             start_seconds=0.0,
@@ -1293,7 +1293,7 @@ class TestMirPersistence:
             metadata={"source": "test"},
         )
 
-        windows = db.list_intent_windows(sample_meeting.id)
+        windows = db.plugins.list_intent_windows(sample_meeting.id)
         assert len(windows) == 1
         window = windows[0]
         assert isinstance(window, IntentWindowSummary)
@@ -1305,7 +1305,7 @@ class TestMirPersistence:
         assert window.metadata["source"] == "test"
 
         # Upsert same window id should refresh values instead of duplicating rows.
-        db.record_intent_window(
+        db.plugins.record_intent_window(
             meeting_id=sample_meeting.id,
             window_id=f"{sample_meeting.id}:w0001",
             start_seconds=30.0,
@@ -1321,7 +1321,7 @@ class TestMirPersistence:
             metadata={"source": "refresh"},
         )
 
-        refreshed = db.list_intent_windows(sample_meeting.id)
+        refreshed = db.plugins.list_intent_windows(sample_meeting.id)
         assert len(refreshed) == 1
         row = refreshed[0]
         assert row.start_seconds == pytest.approx(30.0)
@@ -1338,7 +1338,7 @@ class TestMirPersistence:
         db.meetings.save_meeting(sample_meeting)
         window_id = f"{sample_meeting.id}:w0001"
 
-        db.record_intent_window(
+        db.plugins.record_intent_window(
             meeting_id=sample_meeting.id,
             window_id=window_id,
             start_seconds=0.0,
@@ -1347,7 +1347,7 @@ class TestMirPersistence:
             intent_scores={"architecture": 0.9},
         )
 
-        db.record_plugin_run(
+        db.plugins.record_plugin_run(
             meeting_id=sample_meeting.id,
             window_id=window_id,
             plugin_id="requirements_extractor",
@@ -1361,7 +1361,7 @@ class TestMirPersistence:
         )
 
         # Same idempotency key should upsert, not duplicate.
-        db.record_plugin_run(
+        db.plugins.record_plugin_run(
             meeting_id=sample_meeting.id,
             window_id=window_id,
             plugin_id="requirements_extractor",
@@ -1374,7 +1374,7 @@ class TestMirPersistence:
             deduped=True,
         )
 
-        db.record_plugin_run(
+        db.plugins.record_plugin_run(
             meeting_id=sample_meeting.id,
             window_id=window_id,
             plugin_id="risk_heatmap",
@@ -1387,7 +1387,7 @@ class TestMirPersistence:
             deduped=False,
         )
 
-        runs = db.list_plugin_runs(sample_meeting.id)
+        runs = db.plugins.list_plugin_runs(sample_meeting.id)
         assert len(runs) == 2
         assert all(isinstance(run, PluginRunSummary) for run in runs)
 
@@ -1406,13 +1406,13 @@ class TestMirPersistence:
         assert second.deduped is True
         assert second.output == {"items": 3}
 
-        filtered = db.list_plugin_runs(sample_meeting.id, window_id=window_id)
+        filtered = db.plugins.list_plugin_runs(sample_meeting.id, window_id=window_id)
         assert len(filtered) == 2
 
     def test_record_and_list_artifacts_with_lineage(self, db, sample_meeting):
         db.meetings.save_meeting(sample_meeting)
 
-        db.record_artifact(
+        db.plugins.record_artifact(
             artifact_id="art-1",
             meeting_id=sample_meeting.id,
             artifact_type="requirements",
@@ -1430,7 +1430,7 @@ class TestMirPersistence:
         )
 
         # Upsert should replace body + sources without duplicating record.
-        db.record_artifact(
+        db.plugins.record_artifact(
             artifact_id="art-1",
             meeting_id=sample_meeting.id,
             artifact_type="requirements",
@@ -1447,7 +1447,7 @@ class TestMirPersistence:
             ],
         )
 
-        artifacts = db.list_artifacts(sample_meeting.id)
+        artifacts = db.plugins.list_artifacts(sample_meeting.id)
         assert len(artifacts) == 1
         artifact = artifacts[0]
         assert isinstance(artifact, ArtifactSummary)
@@ -1464,7 +1464,7 @@ class TestMirPersistence:
         db.meetings.save_meeting(sample_meeting)
         idempotency_key = "queue-key-1"
 
-        inserted = db.enqueue_plugin_run_job(
+        inserted = db.plugins.enqueue_plugin_run_job(
             meeting_id=sample_meeting.id,
             window_id=f"{sample_meeting.id}:w0001",
             plugin_id="risk_heatmap",
@@ -1475,7 +1475,7 @@ class TestMirPersistence:
         )
         assert inserted is True
 
-        inserted_again = db.enqueue_plugin_run_job(
+        inserted_again = db.plugins.enqueue_plugin_run_job(
             meeting_id=sample_meeting.id,
             window_id=f"{sample_meeting.id}:w0001",
             plugin_id="risk_heatmap",
@@ -1486,38 +1486,38 @@ class TestMirPersistence:
         )
         assert inserted_again is False
 
-        queued = db.list_plugin_run_jobs(status="queued")
+        queued = db.plugins.list_plugin_run_jobs(status="queued")
         assert len(queued) == 1
         assert queued[0].idempotency_key == idempotency_key
-        loaded = db.get_plugin_run_job(queued[0].id)
+        loaded = db.plugins.get_plugin_run_job(queued[0].id)
         assert loaded is not None
         assert loaded.id == queued[0].id
         assert loaded.status == "queued"
 
-        claimed = db.claim_next_plugin_run_job()
+        claimed = db.plugins.claim_next_plugin_run_job()
         assert claimed is not None
         assert claimed.idempotency_key == idempotency_key
         assert claimed.status == "running"
         assert claimed.attempts == 1
 
-        db.retry_plugin_run_job(
+        db.plugins.retry_plugin_run_job(
             claimed.id,
             error="Transient failure",
             retry_at=datetime.now() - timedelta(seconds=1),
         )
-        claimed_again = db.claim_next_plugin_run_job()
+        claimed_again = db.plugins.claim_next_plugin_run_job()
         assert claimed_again is not None
         assert claimed_again.id == claimed.id
         assert claimed_again.attempts == 2
         assert claimed_again.status == "running"
 
-        db.complete_plugin_run_job(claimed_again.id)
-        assert db.list_plugin_run_jobs(status="all") == []
-        assert db.get_plugin_run_job(claimed_again.id) is None
+        db.plugins.complete_plugin_run_job(claimed_again.id)
+        assert db.plugins.list_plugin_run_jobs(status="all") == []
+        assert db.plugins.get_plugin_run_job(claimed_again.id) is None
 
     def test_plugin_run_job_fail_status(self, db, sample_meeting):
         db.meetings.save_meeting(sample_meeting)
-        db.enqueue_plugin_run_job(
+        db.plugins.enqueue_plugin_run_job(
             meeting_id=sample_meeting.id,
             window_id=f"{sample_meeting.id}:w0002",
             plugin_id="incident_timeline",
@@ -1526,11 +1526,11 @@ class TestMirPersistence:
             idempotency_key="queue-key-fail",
             context={"active_intents": ["incident"]},
         )
-        claimed = db.claim_next_plugin_run_job()
+        claimed = db.plugins.claim_next_plugin_run_job()
         assert claimed is not None
-        db.fail_plugin_run_job(claimed.id, error="Timed out repeatedly")
+        db.plugins.fail_plugin_run_job(claimed.id, error="Timed out repeatedly")
 
-        failed = db.list_plugin_run_jobs(status="failed")
+        failed = db.plugins.list_plugin_run_jobs(status="failed")
         assert len(failed) == 1
         assert failed[0].id == claimed.id
         assert failed[0].status == "failed"
@@ -1539,7 +1539,7 @@ class TestMirPersistence:
     def test_plugin_run_job_summary_reports_queue_telemetry(self, db, sample_meeting):
         db.meetings.save_meeting(sample_meeting)
         for idx in range(1, 5):
-            db.enqueue_plugin_run_job(
+            db.plugins.enqueue_plugin_run_job(
                 meeting_id=sample_meeting.id,
                 window_id=f"{sample_meeting.id}:w{idx:04d}",
                 plugin_id=f"plugin-{idx}",
@@ -1549,19 +1549,19 @@ class TestMirPersistence:
                 context={"active_intents": ["incident"]},
             )
 
-        running = db.claim_next_plugin_run_job()
+        running = db.plugins.claim_next_plugin_run_job()
         assert running is not None
-        failed = db.claim_next_plugin_run_job()
+        failed = db.plugins.claim_next_plugin_run_job()
         assert failed is not None
-        db.fail_plugin_run_job(failed.id, error="Permanent failure")
+        db.plugins.fail_plugin_run_job(failed.id, error="Permanent failure")
 
-        queued = db.list_plugin_run_jobs(status="queued")
+        queued = db.plugins.list_plugin_run_jobs(status="queued")
         assert len(queued) == 2
         scheduled = queued[0]
         retry_at = datetime.now() + timedelta(minutes=5)
-        db.retry_plugin_run_job(scheduled.id, error="Retry later", retry_at=retry_at)
+        db.plugins.retry_plugin_run_job(scheduled.id, error="Retry later", retry_at=retry_at)
 
-        summary = db.get_plugin_run_job_summary()
+        summary = db.plugins.get_plugin_run_job_summary()
         assert summary.total_jobs == 4
         assert summary.queued_jobs == 2
         assert summary.running_jobs == 1
@@ -1572,7 +1572,7 @@ class TestMirPersistence:
 
     def test_claim_next_plugin_run_job_can_include_scheduled(self, db, sample_meeting):
         db.meetings.save_meeting(sample_meeting)
-        db.enqueue_plugin_run_job(
+        db.plugins.enqueue_plugin_run_job(
             meeting_id=sample_meeting.id,
             window_id=f"{sample_meeting.id}:w-scheduled",
             plugin_id="incident_timeline",
@@ -1582,13 +1582,13 @@ class TestMirPersistence:
             context={"active_intents": ["incident"]},
         )
 
-        queued = db.list_plugin_run_jobs(status="queued")
+        queued = db.plugins.list_plugin_run_jobs(status="queued")
         assert len(queued) == 1
         retry_at = datetime.now() + timedelta(minutes=10)
-        db.retry_plugin_run_job(queued[0].id, error="retry later", retry_at=retry_at)
+        db.plugins.retry_plugin_run_job(queued[0].id, error="retry later", retry_at=retry_at)
 
-        assert db.claim_next_plugin_run_job() is None
-        claimed = db.claim_next_plugin_run_job(include_scheduled=True)
+        assert db.plugins.claim_next_plugin_run_job() is None
+        claimed = db.plugins.claim_next_plugin_run_job(include_scheduled=True)
         assert claimed is not None
         assert claimed.id == queued[0].id
         assert claimed.status == "running"
